+++ .github/zz_generated.windows-code-signing.sh
@@ -2,7 +2,7 @@
 
 # DO NOT EDIT. Generated with:
 #
-#    devctl@6.17.0
+#    devctl@6.18.0
 #
 
 APPLICATION=$1