+++ cmd/template/cluster/provider/templates/eks/types.go
@@ -5,8 +5,12 @@ type DefaultAppsConfig struct {
 	Organization string `json:"organization,omitempty"`
 }
 
-type ClusterConfig struct {
+type Global struct {
 	Metadata *Metadata `json:"metadata,omitempty"`
+}
+
+type ClusterConfig struct {
+	Global *Global `json:"global,omitempty"`
 }
 
 type Metadata struct {